新增内容和用户的 demo；完善 sdk 里内容和用户的代码；完善 db 相关的基础代码

+++ app/src/main/java/com/minapp/android/example/database/list/datasource/HorseDataSourceFactory.kt
@@ -6,9 +6,9 @@ import com.minapp.android.example.database.list.ListViewModel
 
 class HorseDataSourceFactory(
     private val viewModel: ListViewModel
-): DataSource.Factory<Long, Horse>() {
+): DataSource.Factory<Int, Horse>() {
 
-    override fun create(): DataSource<Long, Horse> {
+    override fun create(): DataSource<Int, Horse> {
         return HorseDataSource(viewModel)
     }
 }
